# src/main/java/com/example/bankcards/dto/response/UserResponse.java
package com.example.bankcards.dto.response;

import com.example.bankcards.entity.enums.Role;
import java.time.LocalDateTime;
import java.util.UUID;

public record UserResponse(
        UUID id,
        String username,
        String fullName,
        Role role,
        boolean enabled,
        LocalDateTime createdAt,
        LocalDateTime updatedAt
) {
}